Add App component folder and move App files into it
